Support collect external partition stats.
