+++ fe/fe-core/src/main/java/org/apache/doris/analysis/AnalyzeProperties.java
@@ -43,7 +43,7 @@ public class AnalyzeProperties {
     public static final String PROPERTY_ANALYSIS_TYPE = "analysis.type";
     public static final String PROPERTY_PERIOD_SECONDS = "period.seconds";
     public static final String PROPERTY_FORCE_FULL = "force.full";
-    public static final String PROPERTY_PARTITION_COLUMN_FROM_SQL = "partition.column.from.sql";
+    public static final String PROPERTY_EXTERNAL_TABLE_USE_SQL = "external.table.use.sql";
     public static final String PROPERTY_USE_AUTO_ANALYZER = "use.auto.analyzer";
 
     public static final AnalyzeProperties DEFAULT_PROP = new AnalyzeProperties(new HashMap<String, String>() {
@@ -72,7 +72,7 @@ public class AnalyzeProperties {
             .add(PROPERTY_PERIOD_SECONDS)
             .add(PROPERTY_PERIOD_CRON)
             .add(PROPERTY_FORCE_FULL)
-            .add(PROPERTY_PARTITION_COLUMN_FROM_SQL)
+            .add(PROPERTY_EXTERNAL_TABLE_USE_SQL)
             .add(PROPERTY_USE_AUTO_ANALYZER)
             .build();
 
@@ -283,8 +283,8 @@ public class AnalyzeProperties {
         return properties.containsKey(PROPERTY_SAMPLE_ROWS);
     }
 
-    public boolean usingSqlForPartitionColumn() {
-        return properties.containsKey(PROPERTY_PARTITION_COLUMN_FROM_SQL);
+    public boolean usingSqlForExternalTable() {
+        return properties.containsKey(PROPERTY_EXTERNAL_TABLE_USE_SQL);
     }
 
     public String toSQL() {
