+++ fe/fe-core/src/main/java/org/apache/doris/statistics/HMSAnalysisTask.java
@@ -17,7 +17,9 @@
 
 package org.apache.doris.statistics;
 
+import org.apache.doris.catalog.Column;
 import org.apache.doris.common.AnalysisException;
+import org.apache.doris.common.Pair;
 import org.apache.doris.datasource.ExternalTable;
 import org.apache.doris.datasource.hive.HMSExternalTable;
 import org.apache.doris.datasource.hive.HiveMetaStoreCache;
@@ -29,8 +31,11 @@ import org.apache.commons.text.StringSubstitutor;
 import org.apache.logging.log4j.LogManager;
 import org.apache.logging.log4j.Logger;
 
+import java.util.Collections;
 import java.util.HashMap;
+import java.util.List;
 import java.util.Map;
+import java.util.Random;
 import java.util.Set;
 
 public class HMSAnalysisTask extends ExternalAnalysisTask {
@@ -56,31 +61,38 @@ public class HMSAnalysisTask extends ExternalAnalysisTask {
         this.hmsExternalTable = table;
     }
 
-
     @Override
-    protected void getColumnStats() throws Exception {
-        if (info.usingSqlForPartitionColumn) {
-            super.getColumnStats();
+    public void doExecute() throws Exception {
+        if (killed) {
+            return;
+        }
+        if (info.usingSqlForExternalTable) {
+            // If user specify with sql in analyze statement, using sql to collect stats.
+            super.doExecute();
         } else {
+            // By default, using HMS stats and partition info to collect stats.
             try {
+                if (StatisticsUtil.enablePartitionAnalyze() && tbl.isPartitionedTable()) {
+                    throw new RuntimeException("HMS doesn't support fetch partition level stats.");
+                }
                 if (isPartitionColumn()) {
                     getPartitionColumnStats();
                 } else {
                     getHmsColumnStats();
                 }
             } catch (Exception e) {
-                LOG.warn("Failed to collect stats for {}col {} using metadata, "
-                        + "fallback to normal collection",
-                        isPartitionColumn() ? "partition " : "", col.getName(), e);
+                LOG.info("Failed to collect stats for {}col {} using metadata, "
+                        + "fallback to normal collection. Because {}",
+                        isPartitionColumn() ? "partition " : "", col.getName(), e.getMessage());
                 /* retry using sql way! */
-                super.getColumnStats();
+                super.doExecute();
             }
         }
     }
 
     // Collect the partition column stats through HMS metadata.
     // Get all the partition values and calculate the stats based on the values.
-    private void getPartitionColumnStats() throws Exception {
+    private void getPartitionColumnStats() {
         Set<String> partitionNames = hmsExternalTable.getPartitionNames();
         Set<String> ndvPartValues = Sets.newHashSet();
         long numNulls = 0;
@@ -115,7 +127,7 @@ public class HMSAnalysisTask extends ExternalAnalysisTask {
         numNulls = numNulls * count / partitionNames.size();
         int ndv = ndvPartValues.size();
 
-        Map<String, String> params = buildStatsParams("NULL");
+        Map<String, String> params = buildSqlParams();
         params.put("row_count", String.valueOf(count));
         params.put("ndv", String.valueOf(ndv));
         params.put("null_count", String.valueOf(numNulls));
@@ -135,7 +147,7 @@ public class HMSAnalysisTask extends ExternalAnalysisTask {
             count = hmsExternalTable.fetchRowCount();
         }
 
-        Map<String, String> params = buildStatsParams("NULL");
+        Map<String, String> params = buildSqlParams();
         Map<StatsType, String> statsParams = new HashMap<>();
         statsParams.put(StatsType.NDV, "ndv");
         statsParams.put(StatsType.NUM_NULLS, "null_count");
@@ -147,7 +159,7 @@ public class HMSAnalysisTask extends ExternalAnalysisTask {
             throw new AnalysisException("some column stats not available");
         }
 
-        long dataSize = Long.valueOf(params.get("avg_len")) * count;
+        long dataSize = Long.parseLong(params.get("avg_len")) * count;
         params.put("row_count", String.valueOf(count));
         params.put("data_size", String.valueOf(dataSize));
 
@@ -196,5 +208,174 @@ public class HMSAnalysisTask extends ExternalAnalysisTask {
             }
         }
         return value.compareTo(currentMax) > 0 ? value : currentMax;
+    }
+
+    @Override
+    protected void doSample() {
+        StringBuilder sb = new StringBuilder();
+        Map<String, String> params = buildSqlParams();
+        params.put("min", getMinFunction());
+        params.put("max", getMaxFunction());
+        params.put("dataSizeFunction", getDataSizeFunction(col, false));
+        Pair<Double, Long> sampleInfo = getSampleInfo();
+        params.put("scaleFactor", String.valueOf(sampleInfo.first));
+        if (LOG.isDebugEnabled()) {
+            LOG.debug("Will do sample collection for column {}", col.getName());
+        }
+        boolean limitFlag = false;
+        boolean bucketFlag = false;
+        // If sample size is too large, use limit to control the sample size.
+        if (needLimit(sampleInfo.second, sampleInfo.first)) {
+            limitFlag = true;
+            long columnSize = 0;
+            for (Column column : table.getFullSchema()) {
+                columnSize += column.getDataType().getSlotSize();
+            }
+            double targetRows = (double) sampleInfo.second / columnSize;
+            // Estimate the new scaleFactor based on the schema.
+            if (targetRows > StatisticsUtil.getHugeTableSampleRows()) {
+                params.put("limit", "limit " + StatisticsUtil.getHugeTableSampleRows());
+                params.put("scaleFactor",
+                        String.valueOf(sampleInfo.first * targetRows / StatisticsUtil.getHugeTableSampleRows()));
+            }
+        }
+        // Single distribution column is not fit for DUJ1 estimator, use linear estimator.
+        Set<String> distributionColumns = tbl.getDistributionColumnNames();
+        if (distributionColumns.size() == 1 && distributionColumns.contains(col.getName().toLowerCase())) {
+            bucketFlag = true;
+            sb.append(LINEAR_ANALYZE_TEMPLATE);
+            params.put("ndvFunction", "ROUND(NDV(`${colName}`) * ${scaleFactor})");
+            params.put("rowCount", "ROUND(count(1) * ${scaleFactor})");
+        } else {
+            if (col.getType().isStringType()) {
+                sb.append(DUJ1_ANALYZE_STRING_TEMPLATE);
+            } else {
+                sb.append(DUJ1_ANALYZE_TEMPLATE);
+            }
+            params.put("dataSizeFunction", getDataSizeFunction(col, true));
+            params.put("ndvFunction", getNdvFunction("ROUND(SUM(t1.count) * ${scaleFactor})"));
+            params.put("rowCount", "ROUND(SUM(t1.count) * ${scaleFactor})");
+        }
+        LOG.info("Sample for column [{}]. Scale factor [{}], "
+                + "limited [{}], is distribute column [{}]",
+                col.getName(), params.get("scaleFactor"), limitFlag, bucketFlag);
+        StringSubstitutor stringSubstitutor = new StringSubstitutor(params);
+        String sql = stringSubstitutor.replace(sb.toString());
+        runQuery(sql);
+    }
+
+    @Override
+    protected void doFull() throws Exception {
+        if (StatisticsUtil.enablePartitionAnalyze() && tbl.isPartitionedTable()) {
+            doPartitionTable();
+        } else {
+            super.doFull();
+        }
+    }
+
+    @Override
+    protected String getPartitionInfo(String partitionName) {
+        // partitionName is like "date=20230101" for one level partition
+        // and like "date=20230101/hour=12" for two level partition
+        String[] parts = partitionName.split("/");
+        if (parts.length == 0) {
+            throw new RuntimeException("Invalid partition name " + partitionName);
+        }
+        StringBuilder sb = new StringBuilder();
+        sb.append(" WHERE ");
+        for (int i = 0; i < parts.length; i++) {
+            String[] split = parts[i].split("=");
+            if (split.length != 2 || split[0].isEmpty() || split[1].isEmpty()) {
+                throw new RuntimeException("Invalid partition name " + partitionName);
+            }
+            sb.append("`");
+            sb.append(split[0]);
+            sb.append("`");
+            sb.append(" = ");
+            sb.append("'");
+            sb.append(split[1]);
+            sb.append("'");
+            if (i < parts.length - 1) {
+                sb.append(" AND ");
+            }
+        }
+        return sb.toString();
+    }
+
+    protected String getSampleHint() {
+        if (tableSample == null) {
+            return "";
+        }
+        if (tableSample.isPercent()) {
+            return String.format("TABLESAMPLE(%d PERCENT)", tableSample.getSampleValue());
+        } else {
+            return String.format("TABLESAMPLE(%d ROWS)", tableSample.getSampleValue());
+        }
+    }
+
+    /**
+     * Get the pair of sample scale factor and the file size going to sample.
+     * While analyzing, the result of count, null count and data size need to
+     * multiply this scale factor to get more accurate result.
+     * @return Pair of sample scale factor and the file size going to sample.
+     */
+    protected Pair<Double, Long> getSampleInfo() {
+        if (tableSample == null) {
+            return Pair.of(1.0, 0L);
+        }
+        long target;
+        // Get list of all files' size in this HMS table.
+        List<Long> chunkSizes = table.getChunkSizes();
+        Collections.shuffle(chunkSizes, new Random(tableSample.getSeek()));
+        long total = 0;
+        // Calculate the total size of this HMS table.
+        for (long size : chunkSizes) {
+            total += size;
+        }
+        if (total == 0) {
+            return Pair.of(1.0, 0L);
+        }
+        // Calculate the sample target size for percent and rows sample.
+        if (tableSample.isPercent()) {
+            target = total * tableSample.getSampleValue() / 100;
+        } else {
+            int columnSize = 0;
+            for (Column column : table.getFullSchema()) {
+                columnSize += column.getDataType().getSlotSize();
+            }
+            target = columnSize * tableSample.getSampleValue();
+        }
+        // Calculate the actual sample size (cumulate).
+        long cumulate = 0;
+        for (long size : chunkSizes) {
+            cumulate += size;
+            if (cumulate >= target) {
+                break;
+            }
+        }
+        return Pair.of(Math.max(((double) total) / cumulate, 1), cumulate);
+    }
+
+    /**
+     * If the size to sample is larger than LIMIT_SIZE (1GB)
+     * and is much larger (1.2*) than the size user want to sample,
+     * use limit to control the total sample size.
+     * @param sizeToRead The file size to sample.
+     * @param factor sizeToRead * factor = Table total size.
+     * @return True if need to limit.
+     */
+    protected boolean needLimit(long sizeToRead, double factor) {
+        long total = (long) (sizeToRead * factor);
+        long target;
+        if (tableSample.isPercent()) {
+            target = total * tableSample.getSampleValue() / 100;
+        } else {
+            int columnSize = 0;
+            for (Column column : table.getFullSchema()) {
+                columnSize += column.getDataType().getSlotSize();
+            }
+            target = columnSize * tableSample.getSampleValue();
+        }
+        return sizeToRead > LIMIT_SIZE && sizeToRead > target * LIMIT_FACTOR;
     }
 }
